requirements.txt
apache-airflow
apache-airflow-providers-postgres
psycopg2-binary
pandas
apache-airflow-providers-clickhouse